requirements.txt
loguru==0.5.3
matplotlib==3.7.2
numpy==1.25.2
pydantic==1.10.8
python_dateutil==2.8.2
pytz==2023.3
scikit_learn==1.2.2
scipy==1.11.1
statsmodels==0.14.0
torch==2.2.1
wandb==0.15.8
python-dotenv==1.0.1